Move setting definitions to respective classes
