+++ src/main/java/org/opensearch/knn/index/codec/KNNCodecService.java
@@ -5,15 +5,21 @@
 
 package org.opensearch.knn.index.codec;
 
+import org.opensearch.common.settings.Setting;
 import org.opensearch.index.codec.CodecServiceConfig;
 import org.apache.lucene.codecs.Codec;
 import org.opensearch.index.codec.CodecService;
 import org.opensearch.index.mapper.MapperService;
 
+import static org.opensearch.common.settings.Setting.Property.IndexScope;
+
 /**
  * KNNCodecService to inject the right KNNCodec version
  */
 public class KNNCodecService extends CodecService {
+    // Setting to determine if index should use custom codec
+    public static final String KNN_INDEX = "index.knn";
+    public static final Setting<Boolean> IS_KNN_INDEX_SETTING = Setting.boolSetting(KNN_INDEX, false, IndexScope);
 
     private final MapperService mapperService;
