+++ src/main/java/org/opensearch/knn/index/util/IndexHyperParametersUtil.java
@@ -16,7 +16,9 @@ import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.extern.log4j.Log4j2;
 import org.opensearch.Version;
-import org.opensearch.knn.index.KNNSettingsDefinitions;
+
+import static org.opensearch.knn.index.mapper.LegacyFieldMapper.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION;
+import static org.opensearch.knn.index.mapper.LegacyFieldMapper.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH;
 
 /**
  * This class acts as an abstraction to get the default hyperparameter values for different parameters used in the
@@ -48,9 +50,9 @@ public class IndexHyperParametersUtil {
         log.debug(
             "Picking up new values of ef_construction : index version : {}, value: {}",
             indexVersion,
-            KNNSettingsDefinitions.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION
+            INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION
         );
-        return KNNSettingsDefinitions.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION;
+        return INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION;
     }
 
     /**
@@ -72,8 +74,8 @@ public class IndexHyperParametersUtil {
         log.debug(
             "Picking up new values of ef_search : index version : {}, value: {}",
             indexVersion,
-            KNNSettingsDefinitions.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH
+            INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH
         );
-        return KNNSettingsDefinitions.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH;
+        return INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH;
     }
 }
