+++ src/test/java/org/opensearch/knn/index/AdvancedFilteringUseCasesIT.java
@@ -21,6 +21,7 @@ import org.opensearch.common.xcontent.XContentFactory;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.knn.KNNRestTestCase;
 import org.opensearch.knn.NestedKnnDocBuilder;
+import org.opensearch.knn.index.query.KNNWeight;
 import org.opensearch.knn.index.util.KNNEngine;
 
 import java.io.IOException;
@@ -453,7 +454,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
         Assert.assertEquals("For engine " + engine + ", totalSearchHits: ", k, parseTotalSearchHits(response));
         if (KNNEngine.getEngine(engine) == KNNEngine.FAISS) {
             // Update the filter threshold to 0 to ensure that we are hitting ANN Search use case for FAISS
-            updateIndexSettings(INDEX_NAME, Settings.builder().put(KNNSettingsDefinitions.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD, 0));
+            updateIndexSettings(INDEX_NAME, Settings.builder().put(KNNWeight.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD, 0));
             response = EntityUtils.toString(performSearch(INDEX_NAME, query).getEntity());
 
             // Validate number of documents returned as the expected number of documents
