+++ src/test/java/org/opensearch/knn/index/KNNSettingsTests.java
@@ -16,6 +16,7 @@ import org.opensearch.common.settings.Settings;
 import org.opensearch.core.common.unit.ByteSizeValue;
 import org.opensearch.env.Environment;
 import org.opensearch.knn.KNNTestCase;
+import org.opensearch.knn.index.query.KNNWeight;
 import org.opensearch.knn.plugin.KNNPlugin;
 import org.opensearch.node.MockNode;
 import org.opensearch.node.Node;
@@ -32,6 +33,10 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+import static org.opensearch.knn.index.KNNCircuitBreaker.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT;
+import static org.opensearch.knn.index.codec.KNNCodecService.KNN_INDEX;
+import static org.opensearch.knn.index.mapper.LegacyFieldMapper.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH;
+import static org.opensearch.knn.index.mapper.LegacyFieldMapper.KNN_ALGO_PARAM_EF_SEARCH;
 import static org.opensearch.test.NodeRoles.dataNode;
 
 public class KNNSettingsTests extends KNNTestCase {
@@ -41,14 +46,11 @@ public class KNNSettingsTests extends KNNTestCase {
     @SneakyThrows
     public void testGetSettingValueFromConfig() {
         long expectedKNNCircuitBreakerLimit = 13;
-        Node mockNode = createMockNode(
-            Map.of(KNNSettingsDefinitions.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT, "\"" + expectedKNNCircuitBreakerLimit + "kb\"")
-        );
+        Node mockNode = createMockNode(Map.of(KNN_MEMORY_CIRCUIT_BREAKER_LIMIT, "\"" + expectedKNNCircuitBreakerLimit + "kb\""));
         mockNode.start();
         ClusterService clusterService = mockNode.injector().getInstance(ClusterService.class);
         KNNSettings.state().initialize(clusterService);
-        long actualKNNCircuitBreakerLimit = ((ByteSizeValue) KNNSettings.state()
-            .getSettingValue(KNNSettingsDefinitions.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT)).getKb();
+        long actualKNNCircuitBreakerLimit = ((ByteSizeValue) KNNSettings.state().getSettingValue(KNN_MEMORY_CIRCUIT_BREAKER_LIMIT)).getKb();
         mockNode.close();
         assertEquals(expectedKNNCircuitBreakerLimit, actualKNNCircuitBreakerLimit);
         assertWarnings();
@@ -60,12 +62,11 @@ public class KNNSettingsTests extends KNNTestCase {
         mockNode.start();
         ClusterService clusterService = mockNode.injector().getInstance(ClusterService.class);
         KNNSettings.state().initialize(clusterService);
-        long actualKNNCircuitBreakerLimit = ((ByteSizeValue) KNNSettings.state()
-            .getSettingValue(KNNSettingsDefinitions.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT)).getKb();
+        long actualKNNCircuitBreakerLimit = ((ByteSizeValue) KNNSettings.state().getSettingValue(KNN_MEMORY_CIRCUIT_BREAKER_LIMIT)).getKb();
         mockNode.close();
         assertEquals(
-            ((ByteSizeValue) KNNSettingsDefinitions.dynamicCacheSettings.get(KNNSettingsDefinitions.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT)
-                .getDefault(Settings.EMPTY)).getKb(),
+            ((ByteSizeValue) KNNSettingsDefinitions.dynamicCacheSettings.get(KNN_MEMORY_CIRCUIT_BREAKER_LIMIT).getDefault(Settings.EMPTY))
+                .getKb(),
             actualKNNCircuitBreakerLimit
 
         );
@@ -86,7 +87,7 @@ public class KNNSettingsTests extends KNNTestCase {
 
         Integer filteredSearchThreshold = KNNSettings.getFilteredExactSearchThreshold(INDEX_NAME);
         mockNode.close();
-        assertEquals(KNNSettingsDefinitions.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD_DEFAULT_VALUE, filteredSearchThreshold);
+        assertEquals(KNNWeight.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD_DEFAULT_VALUE, filteredSearchThreshold);
         assertWarnings();
     }
 
@@ -102,7 +103,7 @@ public class KNNSettingsTests extends KNNTestCase {
         KNNSettings.state().initialize(clusterService);
 
         final Settings filteredSearchAdvanceSettings = Settings.builder()
-            .put(KNNSettingsDefinitions.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD, userDefinedThreshold)
+            .put(KNNWeight.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD, userDefinedThreshold)
             .build();
 
         mockNode.client()
@@ -115,7 +116,7 @@ public class KNNSettingsTests extends KNNTestCase {
 
         // validate if we are able to set MinValues for the setting
         final Settings filteredSearchAdvanceSettingsWithMinValues = Settings.builder()
-            .put(KNNSettingsDefinitions.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD, userDefinedThresholdMinValue)
+            .put(KNNWeight.ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD, userDefinedThresholdMinValue)
             .build();
 
         mockNode.client()
@@ -143,7 +144,7 @@ public class KNNSettingsTests extends KNNTestCase {
 
         Integer efSearchValue = KNNSettings.getEfSearchParam(INDEX_NAME);
         mockNode.close();
-        assertEquals(KNNSettingsDefinitions.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH, efSearchValue);
+        assertEquals(INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH, efSearchValue);
         assertWarnings();
     }
 
@@ -154,10 +155,7 @@ public class KNNSettingsTests extends KNNTestCase {
         mockNode.start();
         ClusterService clusterService = mockNode.injector().getInstance(ClusterService.class);
         mockNode.client().admin().cluster().state(new ClusterStateRequest()).actionGet();
-        final Settings settings = Settings.builder()
-            .put(KNNSettingsDefinitions.KNN_ALGO_PARAM_EF_SEARCH, userProvidedEfSearch)
-            .put(KNNSettingsDefinitions.KNN_INDEX, true)
-            .build();
+        final Settings settings = Settings.builder().put(KNN_ALGO_PARAM_EF_SEARCH, userProvidedEfSearch).put(KNN_INDEX, true).build();
         mockNode.client().admin().indices().create(new CreateIndexRequest(INDEX_NAME, settings)).actionGet();
         KNNSettings.state().initialize(clusterService);
